Resolved smoke test issue
