--- src/smokeTest/groovy/uk/gov/hmcts/reform/em/hrs/smoke/config/SmokeTestContextConfiguration.groovy
@@ -1,11 +1,0 @@
-package uk.gov.hmcts.reform.em.hrs.smoke.config
-
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.PropertySource
-
-@Configuration
-@PropertySource("classpath:application-smoke-test.yaml")
-@ComponentScan("uk.gov.hmcts.reform.em.hrs.smoke")
-class SmokeTestContextConfiguration {
-}
