--- src/smokeTest/groovy/uk/gov/hmcts/reform/em/hrs/smoke/utilities/Classifications.groovy
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.em.hrs.smoke.utilities
-
-enum Classifications {
-    PUBLIC,PRIVATE,RESTRICTED
-}
