--- src/smokeTest/groovy/uk/gov/hmcts/reform/em/hrs/smoke/utilities/V1MediaTypes.groovy
@@ -1,27 +1,0 @@
-package uk.gov.hmcts.reform.em.hrs.smoke.utilities
-
-class V1MediaTypes {
-
-    public static final String V1_HAL_DOCUMENT_AND_METADATA_COLLECTION_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.document-and-metadata-collection.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_HAL_DOCUMENT_PAGE_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.document-page.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_HAL_DOCUMENT_COLLECTION_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.document-collection.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_HAL_DOCUMENT_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.document.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_HAL_FOLDER_MEDIA_TYPE_VALUE="application/vnd.uk.gov.hmcts.dm.folder.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_HAL_DOCUMENT_CONTENT_VERSION_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.documentContentVersion.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_HAL_AUDIT_ENTRY_MEDIA_TYPE_VALUE="application/vnd.uk.gov.hmcts.dm.auditEntry.v1+hal+json;charset=UTF-8"
-
-    public static final String V1_DOCUMENT_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.document.v1+json;charset=UTF-8"
-
-    public static final String V1_FOLDER_MEDIA_TYPE_VALUE="application/vnd.uk.gov.hmcts.dm.folder.v1+json;charset=UTF-8"
-
-    public static final String V1_DOCUMENT_CONTENT_VERSION_MEDIA_TYPE_VALUE ="application/vnd.uk.gov.hmcts.dm.documentContentVersion.v1+json;charset=UTF-8"
-
-    public static final String V1_AUDIT_ENTRY_MEDIA_TYPE_VALUE="application/vnd.uk.gov.hmcts.dm.auditEntry.v1+json;charset=UTF-8"
-
-}
