--- src/smokeTest/resources/application-smoke-test.yaml
@@ -1,16 +1,0 @@
-base-urls.idam-user=${IDAM_API_URL:http://localhost:4501}
-base-urls.idam-s2s=${S2S_URL:http://localhost:4502}
-base-urls.dm-store=${TEST_URL:http://localhost:4603}
-
-login.username=${TEST_USERNAME:test@test.com}
-login.password=${TEST_PASSWORD:password}
-login.token=${TEST_TOKEN:token}
-
-toggle.metadatasearchendpoint=${ENABLE_METADATA_SEARCH:true}
-toggle.documentandmetadatauploadendpoint=${ENABLE_DOCUMENT_AND_METADATA_UPLOAD:true}
-toggle.folderendpoint=${ENABLE_FOLDER_API:true}
-toggle.includeidamhealth=${ENABLE_IDAM_HEALTH_CHECK:false}
-toggle.deleteenabled=${ENABLE_DELETE:true}
-toggle.ttl=${ENABLE_TTL:true}
-toggle.thumbnail=${ENABLE_THUMBNAIL:true}
-toggle.testing=${ENABLE_TESTING:true}
